Refactored code.| #1947
